Move queryDefaultPageData to pages folder
