+++ functions/wordpress/postTypes/getFrontendPage.js
@@ -1,8 +1,8 @@
 import formatDefaultSeoData from '@/functions/formatDefaultSeoData'
 import {initializeWpApollo} from '@/lib/wordpress/connector'
 import getMenus from '@/lib/wordpress/menus/getMenus'
+import queryDefaultPageData from '@/lib/wordpress/pages/queryDefaultPageData'
 import frontendPageSeo from '@/lib/wordpress/_config/frontendPageSeo'
-import queryDefaultPageData from '@/lib/wordpress/_global/queryDefaultPageData'
 
 /**
  * Retrieve data for Frontend-only route (i.e., page does not exist in WordPress).
