+++ lib/wordpress/pages/queryDefaultPageData.js
